+++ examples/basic_table.rb
@@ -15,10 +15,10 @@ data = [
 window('Animal sounds', 300, 200) {
   horizontal_box {
     table {
-#       text_column('Animal')
-#       text_column('Description')
-#
-#       cell_values data
+      text_column('Animal')
+      text_column('Description')
+
+      cell_rows data
     }
   }